Be less aggressive with whitespace trimming
Why:
To keep things tidy the entire process string was having
leading/trailing whitespace stripped. This was lazily done to remove the
trailing whitespace but had the undesired effect of messing up
indentation on the first line of the replacement.

This change addresses the need by:
Only remove trailing whitespace.